pyproject.toml
[build-system]
requires = ["setuptools >= 64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clproc"
version = "2024.02.08"
authors = [
  { name="Michel Albert", email="michel.albert@post.lu" },
]
description = "Process a changelog.in file into various outputs"
readme = "README.rst"
license = { text="MIT License" }
requires-python = ">=3.7"
dependencies = [
    "packaging",
    "python-dateutil",
    "pyyaml",
    "tomli",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
clproc = "clproc.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cache",
    "pytest-coverage",
]
dev = [
    "recommonmark",
    "sphinx",
    "sphinx-rtd-theme",
    "types-PyYAML",
    "types-python-dateutil",
]

[project.urls]
"Repository" = "https://github.com/post-luxembourg/clproc"

[tool.black]
line_length = 80

[tool.isort]
line_length = 80
force_grid_wrap = 0
virtual_env = "env"
profile = "black"

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_", "id", "ip"]

[tool.pylint.format]
max-line-length = "80"
# Ignore RST external links (f.ex.: ".. _thelink: https://example.com")
ignore-long-lines = '^\s*(\.\. _\w+: )?https?://.*'

[tool.coverage.run]
relative_files = true
source = ["clproc"]
branch = true
omit = [
    "env/*",
    "tests/*",
]

[tool.coverage.report]
skip_covered = true
skip_empty = true
sort = "-cover"

[tool.coverage.html]
show_contexts = true

[tool.pyright]
# strict = ["clproc"]
exclude = ["fabfile.py", "env"]
ignore = ["tests"]
venvPath = "."
venv = "env"

[tool.mypy]
strict = true
